sql: move schema change backfill logic to sql/backfill
The distsql backfill calls the sql/backfill code.

The future plan is to also call backfill within a transaction
that runs schema changes in the same transaction as a
CREATE TABLE. Since the CREATE TABLE in a transaction
is not visible to other transactions until the transaction
commits, the schema change can run the backfill without
incrementing the version of the table descriptor and running
the schema change state machine.

related to #24626

Release note: None
