+++ pkg/sql/distsqlrun/indexbackfiller.go
@@ -19,11 +19,9 @@ import (
 
 	"github.com/cockroachdb/cockroach/pkg/internal/client"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
+	"github.com/cockroachdb/cockroach/pkg/sql/backfill"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
-	"github.com/cockroachdb/cockroach/pkg/sql/scrub"
-	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sqlbase"
-	"github.com/cockroachdb/cockroach/pkg/util"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/cockroach/pkg/util/tracing"
@@ -33,21 +31,11 @@ import (
 type indexBackfiller struct {
 	backfiller
 
-	// colIdxMap maps ColumnIDs to indices into desc.Columns and desc.Mutations.
-	colIdxMap map[sqlbase.ColumnID]int
-
-	types   []sqlbase.ColumnType
-	rowVals tree.Datums
-	da      sqlbase.DatumAlloc
+	backfill.IndexBackfiller
 }
 
 var _ Processor = &indexBackfiller{}
 var _ chunkBackfiller = &indexBackfiller{}
-
-// IndexMutationFilter is a filter that allows mutations that add indexes.
-func IndexMutationFilter(m sqlbase.DescriptorMutation) bool {
-	return m.GetIndex() != nil && m.Direction == sqlbase.DescriptorMutation_ADD
-}
 
 func newIndexBackfiller(
 	flowCtx *FlowCtx, spec BackfillerSpec, post *PostProcessSpec, output RowReceiver,
@@ -55,7 +43,7 @@ func newIndexBackfiller(
 	ib := &indexBackfiller{
 		backfiller: backfiller{
 			name:    "Index",
-			filter:  IndexMutationFilter,
+			filter:  backfill.IndexMutationFilter,
 			flowCtx: flowCtx,
 			output:  output,
 			spec:    spec,
@@ -63,63 +51,11 @@ func newIndexBackfiller(
 	}
 	ib.backfiller.chunkBackfiller = ib
 
-	if err := ib.init(); err != nil {
+	if err := ib.IndexBackfiller.Init(ib.spec.Table); err != nil {
 		return nil, err
 	}
 
 	return ib, nil
-}
-
-func (ib *indexBackfiller) init() error {
-	desc := ib.spec.Table
-	numCols := len(desc.Columns)
-
-	cols := desc.Columns
-	if len(desc.Mutations) > 0 {
-		cols = make([]sqlbase.ColumnDescriptor, 0, numCols+len(desc.Mutations))
-		cols = append(cols, desc.Columns...)
-		for _, m := range desc.Mutations {
-			if column := m.GetColumn(); column != nil {
-				cols = append(cols, *column)
-			}
-		}
-	}
-	ib.types = make([]sqlbase.ColumnType, len(cols))
-	for i := range cols {
-		ib.types[i] = cols[i].Type
-	}
-
-	ib.colIdxMap = make(map[sqlbase.ColumnID]int, len(cols))
-	for i, c := range cols {
-		ib.colIdxMap[c.ID] = i
-	}
-
-	var valNeededForCol util.FastIntSet
-	mutationID := desc.Mutations[0].MutationID
-	for _, m := range desc.Mutations {
-		if m.MutationID != mutationID {
-			break
-		}
-		if IndexMutationFilter(m) {
-			idx := m.GetIndex()
-			for i, col := range cols {
-				if idx.ContainsColumnID(col.ID) {
-					valNeededForCol.Add(i)
-				}
-			}
-		}
-	}
-
-	tableArgs := sqlbase.RowFetcherTableArgs{
-		Desc:            &desc,
-		Index:           &desc.PrimaryIndex,
-		ColIdxMap:       ib.colIdxMap,
-		Cols:            cols,
-		ValNeededForCol: valNeededForCol,
-	}
-	return ib.fetcher.Init(
-		false /* reverse */, false /* returnRangeInfo */, false /* isCheck */, &ib.alloc, tableArgs,
-	)
 }
 
 func (ib *indexBackfiller) runChunk(
@@ -145,72 +81,14 @@ func (ib *indexBackfiller) runChunk(
 	for i, m := range mutations {
 		added[i] = *m.GetIndex()
 	}
-	secondaryIndexEntries := make([]sqlbase.IndexEntry, len(mutations))
 
-	buildIndexEntries := func(ctx context.Context, txn *client.Txn) ([]sqlbase.IndexEntry, error) {
-		entries := make([]sqlbase.IndexEntry, 0, chunkSize*int64(len(added)))
-
-		// Get the next set of rows.
-		//
-		// Running the scan and applying the changes in many transactions
-		// is fine because the schema change is in the correct state to
-		// handle intermediate OLTP commands which delete and add values
-		// during the scan. Index entries in the new index are being
-		// populated and deleted by the OLTP commands but not otherwise
-		// read or used
-		if err := ib.fetcher.StartScan(
-			ctx, txn, []roachpb.Span{sp}, true /* limitBatches */, chunkSize, false, /* traceKV */
-		); err != nil {
-			log.Errorf(ctx, "scan error: %s", err)
-			return nil, err
-		}
-
-		for i := int64(0); i < chunkSize; i++ {
-			encRow, _, _, err := ib.fetcher.NextRow(ctx)
-			if err != nil {
-				err = scrub.UnwrapScrubError(err)
-				return nil, err
-			}
-			if encRow == nil {
-				break
-			}
-			if len(ib.rowVals) == 0 {
-				ib.rowVals = make(tree.Datums, len(encRow))
-			}
-			if err := sqlbase.EncDatumRowToDatums(ib.types, ib.rowVals, encRow, &ib.da); err != nil {
-				return nil, err
-			}
-
-			// We're resetting the length of this slice for variable length indexes such as inverted
-			// indexes which can append entries to the end of the slice. If we don't do this, then everything
-			// EncodeSecondaryIndexes appends to secondaryIndexEntries for a row, would stay in the slice for
-			// subsequent rows and we would then have duplicates in entries on output.
-			secondaryIndexEntries = secondaryIndexEntries[:len(mutations)]
-			if secondaryIndexEntries, err = sqlbase.EncodeSecondaryIndexes(
-				&ib.spec.Table, added, ib.colIdxMap,
-				ib.rowVals, secondaryIndexEntries); err != nil {
-				return nil, err
-			}
-			entries = append(entries, secondaryIndexEntries...)
-		}
-		return entries, nil
-	}
-
+	var key roachpb.Key
 	transactionalChunk := func(ctx context.Context) error {
 		return ib.flowCtx.clientDB.Txn(ctx, func(ctx context.Context, txn *client.Txn) error {
-			entries, err := buildIndexEntries(ctx, txn)
-			if err != nil {
-				return err
-			}
-			batch := txn.NewBatch()
-
-			for _, entry := range entries {
-				batch.InitPut(entry.Key, &entry.Value, false /* failOnTombstones */)
-			}
-			if err := txn.CommitInBatch(ctx, batch); err != nil {
-				return ConvertBackfillError(ctx, &ib.spec.Table, batch)
-			}
-			return nil
+			var err error
+			key, err = ib.RunIndexBackfillChunk(
+				ctx, txn, ib.spec.Table, sp, chunkSize, true /*alsoCommit*/)
+			return err
 		})
 	}
 
@@ -231,7 +109,7 @@ func (ib *indexBackfiller) runChunk(
 		txn.SetFixedTimestamp(ctx, readAsOf)
 
 		var err error
-		entries, err = buildIndexEntries(ctx, txn)
+		entries, key, err = ib.BuildIndexEntriesChunk(ctx, txn, ib.spec.Table, sp, chunkSize)
 		return err
 	}); err != nil {
 		return nil, err
@@ -283,5 +161,5 @@ func (ib *indexBackfiller) runChunk(
 		}
 	}
 
-	return ib.fetcher.Key(), nil
+	return key, nil
 }
